Remove float64 in softmax, which is unnecessary

+++ generate.py
@@ -85,7 +85,7 @@ def generate(model, prompt, steps=128, gen_length=128, block_length=128, tempera
             x0 = torch.argmax(logits_with_noise, dim=-1) # b, l
 
             if remasking == 'low_confidence':
-                p = F.softmax(logits.to(torch.float64), dim=-1)
+                p = F.softmax(logits, dim=-1)
                 x0_p = torch.squeeze(
                     torch.gather(p, dim=-1, index=torch.unsqueeze(x0, -1)), -1) # b, l
             elif remasking == 'random':
